requirements.txt
matplotlib==3.8.0
networkx==3.1
numpy==1.26.0
numpy_stl==3.0.1
pandas==2.1.1
scikit_learn==1.3.0
torch==2.0.1
